refactor, and add project level merge
